Release v3.1.0

+++ lib/src/common/models/chart_model/chart_data_model.dart
@@ -4,6 +4,6 @@ class ChartDataModel {
   ChartDataModel({this.x = '', required this.y, required this.color});
 
   final String x;
-  final int y;
+  final double y;
   final Color color;
 }
